Create separate proxy file and remove unused code
